refactor(depth-chart): move schema to its owning feature folder
depth-chart.schema.ts lived in server/features/players/ but was
consumed primarily by depth-chart.publisher, roster.repository, and
roster-transaction — not by players. Location didn't match ownership,
which made "who owns this table" ambiguous when reading cross-feature
imports.

Move server/features/players/depth-chart.schema.ts →
server/features/depth-chart/depth-chart.schema.ts and update the five
import sites (db/schema.ts, depth-chart.publisher + test,
roster.repository + test, roster-transaction test).

Cross-feature joins that reference depthChartEntries inside larger
Drizzle queries (roster's active-roster query, for instance) are left
as-is — schema location signals ownership, it doesn't forbid joins.

No behavior change. All 708 server tests pass.
